\name{drinks}
\alias{drinks}
\docType{data}
\title{
soft-drinks data}
\description{
A one-year categorical time series of daily sales of soft-drinks.}
\usage{data(drinks)}
\format{
  A five-variate time series of sale levels of soft-drinks on 269 days
  \describe{
    \item{\code{lemon.tea}}{A factor with levels: 1 = low, 2 = medium, 3 = high}
    \item{\code{orange.juice}}{A factor with levels: 1 = low, 2 = medium, 3 = high}
    \item{\code{apple.juice}}{A factor with levels: 1 = low, 2 = medium, 3 = high}
    \item{\code{mint.tea}}{A factor with levels: 1 = low, 2 = medium, 3 = high}
    \item{\code{soya.milk}}{A factor with levels: 1 = low, 2 = medium, 3 = high}
  }
}
\references{
Colombi R, Giordano S (2011) Lumpability for discrete hidden Markov models. Advances in Statistical Analysis, 95(3), 293-311.}
\examples{
data(drinks)}
\keyword{datasets}
