\name{hmmm-package}
\alias{hmmm-package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{package hmmm}
\description{Functions for specifying and fitting marginal models for 
contingency tables proposed by Bergsma and Rudas (2002)
 here called hierarchical multinomial marginal models (hmmm) and their extensions presented
 by Bartolucci et al (2007); multinomial Poisson homogeneous (mph) models and homogeneous linear predictor (hlp) models for contingency 
tables proposed by Lang (2004, 2005); hidden Markov models where the distribution of the observed variables is described by a marginal model. 
Inequality constraints on the parameters are allowed and can be tested.}

\author{Roberto Colombi, Sabrina Giordano and Manuela Cazzaro. Joseph B. Lang is the author
of functions `num.deriv.fct', `create.U' for mph models.}
\references{
Bergsma WP, Rudas T (2002) 
Marginal models for categorical data. The Annals of Statistics, 30, 140-159

Bartolucci F, Colombi R, Forcina A (2007) An extended class of marginal link functions for modelling contingency 
tables by equality and inequality constraints. Statistica Sinica, 17, 691-711

Lang  JB (2004) Multinomial Poisson homogeneous models for contingency tables. The Annals of Statistics, 32, 340-383

Lang JB (2005) Homogeneous linear predictor models for contingency tables. Journal of the American Statistical Association, 100, 121-134.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{package}
\docType{package}