\name{relpolbirth}
\alias{relpolbirth}
\docType{data}
\title{religion, political orientation and teenage birth control data}
\description{Data on religion, political orientation and opinion on teenage birth control of a sample of 911 U.S. citizens.}
\usage{data(relpolbirth)}
\format{A data frame whose columns contain:
\describe{\item{\code{Religion}}{A factor with levels: \code{Protestant}, \code{Catholic}, \code{None}}
\item{\code{Politics}}{A factor with levels: \code{Extremely liberal}, \code{Liberal}, \code{Slightly liberal}, \code{Moderate},
\code{Slightly conservative}, \code{Conservative}, \code{Extremely conservative}}
\item{\code{Birthcontrol}}{A factor with levels: \code{Strongly agree}, \code{Agree}, \code{Disagree}, \code{Strongly disagree}} 
\item{\code{Freq}}{A numeric vector of frequencies}}}
\source{General Social Survey, 1993}
\references{Bergsma W, Croon M, Hagenaars JA (2009) Marginal models for dependent, clustered, and longitudinal categorical data. Springer.}
\examples{data(relpolbirth)}
\keyword{datasets}
